refactor(deploy): change id type => Long

+++ src/main/java/chen/binflare/dao/UserCredentialDAO.java
@@ -2,7 +2,5 @@ package chen.binflare.dao;
 
 import chen.binflare.entity.UserCredentialDO;
 
-import java.math.BigInteger;
-
-public interface UserCredentialDAO extends BaseRepository<UserCredentialDO, BigInteger> {
+public interface UserCredentialDAO extends BaseRepository<UserCredentialDO, Long> {
 }
