+++ src/main/java/chen/binflare/dao/UserDAO.java
@@ -2,7 +2,5 @@ package chen.binflare.dao;
 
 import chen.binflare.entity.UserDO;
 
-import java.math.BigInteger;
-
-public interface UserDAO extends BaseRepository<UserDO, BigInteger> {
+public interface UserDAO extends BaseRepository<UserDO, Long> {
 }
